Refactor Engine in to multiple files
